[explorer/rest]: refactor database test utils
